build(book-api): upgrade Spring Boot to 4.0.5 and modernise backend code
- bump spring-boot-starter-parent from 4.0.3 to 4.0.5
- replace String.format calls with String.formatted for consistency
- change countByRole and countAdmins return types from int to long
- remove redundant @Repository annotations from JPA repository interfaces
- make getUsers and getBooks helper methods static in DatabaseInitializer
- inject ObjectMapper via @Autowired in tests instead of constructing manually
- update ObjectMapper import to tools.jackson.databind in test classes
- fix countAdmins mock return values to use long literals in tests
- add jq presence check with helpful error message to test-endpoints.sh
- update brace-expansion and picomatch in package-lock.json
- update AGENTS.md to reflect Spring Boot 4.0.5

+++ book-api/src/main/java/com/ivanfranchin/bookapi/user/UserService.java
@@ -9,7 +9,7 @@ public interface UserService {
 
     long countUsers();
 
-    int countAdmins();
+    long countAdmins();
 
     Optional<User> getUserByUsername(String username);
